Study Home - Step Count Goal & Daily Step Count Display (#7)
# Study Home - Step Count Goal & Daily Step Count Display

## :gear: Release Notes 
- Adds a study dashboard section for all enrolled studies.
- Allows a user to define a step count goal & display their current
daily steps.

## :pencil: Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md):
- [x] I agree to follow the [Code of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md).

+++ StudyApplication/StudyApplicationStandard.swift
@@ -68,6 +68,11 @@ actor StudyApplicationStandard: Standard, EnvironmentAccessible, HealthKitConstr
 
 
     func add(sample: HKSample) async {
+        guard !FeatureFlags.disableFirebase else {
+            logger.debug("Firebase disabled - would upload HKSample: \(sample)")
+            return
+        }
+        
         do {
             try await healthKitDocument(id: sample.id).setData(from: sample.resource)
         } catch {
@@ -76,6 +81,11 @@ actor StudyApplicationStandard: Standard, EnvironmentAccessible, HealthKitConstr
     }
     
     func remove(sample: HKDeletedObject) async {
+        guard !FeatureFlags.disableFirebase else {
+            logger.debug("Firebase disabled - would remove HKDeletedObject: \(sample)")
+            return
+        }
+        
         do {
             try await healthKitDocument(id: sample.uuid).delete()
         } catch {
@@ -84,6 +94,11 @@ actor StudyApplicationStandard: Standard, EnvironmentAccessible, HealthKitConstr
     }
     
     func add(response: ModelsR4.QuestionnaireResponse) async {
+        guard !FeatureFlags.disableFirebase else {
+            logger.debug("Firebase disabled - would upload questionnaire response: \(response.description)")
+            return
+        }
+        
         let id = response.identifier?.value?.value?.string ?? UUID().uuidString
         
         do {
